Páginas de crear evento con vistas para escritorio y movil
